feat: integrando com serverless
