+++ package.json
@@ -28,6 +28,8 @@
     "@types/pg": "^8.11.11",
     "@types/sequelize": "^4.28.20",
     "nodemon": "^3.1.9",
+    "serverless-http": "^3.2.0",
+    "serverless-offline": "^14.4.0",
     "ts-node": "^10.9.2",
     "typescript": "^5.7.3"
   }
